requirements.txt
absl-py==0.11.0
apturl==0.5.2
astunparse==1.6.3
atari-py==0.2.6
attrs==20.3.0
baselines==0.1.6
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudpickle==1.2.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
gym==0.15.7
h5py==2.10.0
httplib2==0.14.0
idna==2.8
iniconfig==1.1.1
joblib==1.0.1
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.0
matplotlib==3.3.4
monotonic==1.5
netifaces==0.10.4
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pluggy==0.13.1
protobuf==3.15.5
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
pyglet==1.5.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==6.2.2
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rl-algs==0.0.0
rsa==4.7.2
scipy==1.4.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
systemd-python==234
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-gpu==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tf-slim==1.1.0
toml==0.10.2
tqdm==4.59.0
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
